+++ AheadOfTheCurve.lua
@@ -1,13 +1,15 @@
 local AheadOfTheCurve = LibStub('AceAddon-3.0'):NewAddon('AheadOfTheCurve', 'AceConsole-3.0', 'AceHook-3.0', 'AceEvent-3.0')
 local AceConfigRegistry = LibStub('AceConfigRegistry-3.0')
-local container
-local checkButton
+local keyStoneLink
 local defaults = {
     global = {
         enable = {
             addon = true,
             override = false,
-            whisper = false
+            whispers = {
+                achievement = false,
+                keystone = false
+            }
         },
         minimap = {
             hide = true
@@ -32,19 +34,19 @@ function AheadOfTheCurve:OnInitialize()
         {
             ids = {413, 414, 468},
             name = 'Emerald Nightmare',
-            achievements = {11191, 11194, 10820},
+            achievements = {11191, 10827, 11194, 10820},
             highestCompleted = nil,
         },
         {
             ids = {456, 457},
             name = 'Trial of Valor',
-            achievements = {11580, 11581, 11394},
+            achievements = {11580, 11398, 11581, 11394},
             highestCompleted = nil,
         },
         {
             ids = {415, 416},
             name = 'The Nighthold',
-            achievements = {11192, 11195, 10839},
+            achievements = {11192, 10850, 11195, 10839},
             highestCompleted = nil,
         },
         {
@@ -56,7 +58,7 @@ function AheadOfTheCurve:OnInitialize()
         {
             ids = {459, 460, 461, 462, 463, 464, 465, 466, 467},
             name = 'Mythic Plus',
-            achievements = {11224, 11162, 11185, 11184},
+            achievements = {11162, 11185, 11184},
             highestCompleted = nil,
         },
         {
@@ -75,12 +77,13 @@ function AheadOfTheCurve:OnEnable()
     self:RegisterEvent('ACHIEVEMENT_SEARCH_UPDATED')
     self:RegisterEvent('ACHIEVEMENT_EARNED')
     self:GetHighestDefaultAchievement()
+    self:ScanBags()
     
-    if not container then
-        container = CreateFrame('Frame', 'AheadOfTheCurveDialog', LFGListApplicationDialog)
-        container:SetSize(306, 50)
-        container:SetPoint('BOTTOM', 0, -55)
-        container:SetBackdrop({
+    if not self.container then
+        self.container = CreateFrame('Frame', 'AheadOfTheCurveDialog', LFGListApplicationDialog)
+        self.container:SetSize(306, 50)
+        self.container:SetPoint('BOTTOM', 0, -55)
+        self.container:SetBackdrop({
             bgFile = 'Interface\\DialogFrame\\UI-DialogBox-Background',
             edgeFile = 'Interface\\DialogFrame\\UI-DialogBox-Border',
             tile = true, tileSize = 32, edgeSize = 32,
@@ -88,25 +91,35 @@ function AheadOfTheCurve:OnEnable()
         })
     end
     
-    if not checkButton then
-        checkButton = CreateFrame('CheckButton', 'AheadOfTheCurveCheckBox', container, 'UICheckButtonTemplate')
-        checkButton:SetSize(24, 24)
-        checkButton:SetPoint('CENTER', -85, 0)
-        checkButton:SetChecked(self.db.global.enable.whisper)
-        checkButton.text:SetText('Use Ahead of the Curve Whisper')
-        checkButton.text:SetWidth(175)
+    if not self.checkButtonAchievement then
+        self.checkButtonAchievement = CreateFrame('CheckButton', 'AheadOfTheCurveCheckBoxAchievement', self.container, 'UICheckButtonTemplate')
+        self.checkButtonAchievement:SetSize(24, 24)
+        self.checkButtonAchievement:SetPoint('CENTER', -85, 0)
+        self.checkButtonAchievement:SetChecked(self.db.global.enable.whispers.achievement)
+        self.checkButtonAchievement.text:SetText('Use Ahead of the Curve Whisper')
+        self.checkButtonAchievement.text:SetWidth(175)
+    end
+
+    if not self.checkButtonKeystone then
+        self.checkButtonKeystone = CreateFrame('CheckButton', 'AheadOfTheCurveCheckBoxKeystone', self.container, 'UICheckButtonTemplate')
+        self.checkButtonKeystone:SetSize(24, 24)
+        self.checkButtonKeystone:SetPoint('CENTER', -85, -11)
+        self.checkButtonKeystone:SetChecked(self.db.global.enable.whispers.keystone)
+        self.checkButtonKeystone.text:SetText('Link Mythic Plus Keystone')
+        self.checkButtonKeystone.text:SetWidth(145)
     end
     
-    container:Show()
-    checkButton:Show()
+    self.container:Show()
+    self.checkButtonAchievement:Show()
+    self.checkButtonKeystone:Hide()
 end
 
 function AheadOfTheCurve:OnDisable()
     self:UnhookAll()
     self:UnregisterEvent('ACHIEVEMENT_SEARCH_UPDATED')
     self:UnregisterEvent('ACHIEVEMENT_EARNED')
-    container:Hide()
-    checkButton:Hide()
+    self.container:Hide()
+    self.checkButtonAchievement:Hide()
 end
 
 function AheadOfTheCurve:ACHIEVEMENT_SEARCH_UPDATED()
@@ -161,24 +174,37 @@ function AheadOfTheCurve:GetLFGCategory(findAGroupButton)
     local category = findAGroupButton:GetParent().selectedCategory;
     
     if category ~= 2 or category ~= 3 then
-        container:Hide()
-        checkButton:Hide()
+        self.container:Hide()
+        self.checkButtonAchievement:Hide()
     else
-        container:Show()
-        checkButton:Show()
+        self.container:Show()
+        self.checkButtonAchievement:Show()
     end
 end
 
 function AheadOfTheCurve:GetLFGInstance(signUpButton)
     local _, instanceId = C_LFGList.GetSearchResultInfo(signUpButton:GetParent().selectedResult)
     local highestCompleted = self:GetLFGAchievement(instanceId)
+    local isMythicDungeon = self:IsMythicDungeon(instanceId)
     
     if not highestCompleted and not self.db.global.enable.override or raidId == self.instances[4].ids[1] then
-        container:Hide()
-        checkButton:Hide()
+        self.container:Hide()
+        self.checkButtonAchievement:Hide()
     else
-        container:Show()
-        checkButton:Show()
+        self.container:Show()
+        self.checkButtonAchievement:Show()
+        self.checkButtonAchievement:SetChecked(self.db.global.enable.whispers.achievement)      
+    end
+
+    if isMythicDungeon and keyStoneLink then
+        self:ShowMythicPlusOption(highestCompleted)
+        self.checkButtonKeystone:SetChecked(self.db.global.enable.whispers.keystone)
+    else
+        self.container:SetPoint('BOTTOM', 0, -55)
+        self.container:SetSize(306, 50)
+        self.checkButtonAchievement:SetPoint('CENTER', -85, 0)
+        self.checkButtonKeystone:Hide()
+        self.checkButtonKeystone:SetChecked(false)
     end
 end
 
@@ -195,7 +221,7 @@ end
 function AheadOfTheCurve:SendWhisper(signUpButton)
     local _, instanceId, _, _, _, _, _, _, _, _, _, _, leaderName = C_LFGList.GetSearchResultInfo(signUpButton:GetParent().resultID)
     
-    if checkButton:GetChecked() or instanceId == self.instances[4].ids[1] then
+    if self.checkButtonAchievement:GetChecked() or self.checkButtonKeystone:GetChecked() or instanceId == self.instances[4].ids[1] then
         local achievementId
         
         if self.db.global.enable.override and self.db.global.overrideAchievement then
@@ -205,11 +231,63 @@ function AheadOfTheCurve:SendWhisper(signUpButton)
         end
         
         if achievementId then
-            local success = pcall(SendChatMessage, GetAchievementLink(achievementId), 'WHISPER', nil, leaderName)
+            local success = pcall(SendChatMessage, self:BuildWhisper(achievementId), 'WHISPER', nil, leaderName)
             
             if not success then
                 self:Print('There was an error sending your whisper.')
             end
         end
     end
+end
+
+function AheadOfTheCurve:BuildWhisper(achievementId)
+    if self.checkButtonAchievement:GetChecked() and (self.checkButtonKeystone:GetChecked() and keyStoneLink) then
+        return GetAchievementLink(achievementId) .. keyStoneLink
+    end
+
+    if self.checkButtonAchievement:GetChecked() then
+        return GetAchievementLink(achievementId)
+    end
+
+    if self.checkButtonKeystone:GetChecked() and keyStoneLink then
+        return keyStoneLink
+    end
+end
+
+function AheadOfTheCurve:ScanBags()
+    for bag = 0, 4 do
+        local numSlots = GetContainerNumSlots(bag)
+
+        for slot = 1, numSlots do
+            local _, _, _, _, _, _, link, _, _, itemId = GetContainerItemInfo(bag, slot)
+
+            if itemId == 138019 then
+                keyStoneLink = link
+                break
+            end
+        end
+    end
+end
+
+function AheadOfTheCurve:IsMythicDungeon(instanceId)
+    for _, id in pairs(self.instances[5].ids) do
+        if id == instanceId then
+            return true
+        end
+    end
+end
+
+function AheadOfTheCurve:ShowMythicPlusOption(highestCompleted)
+     if not highestCompleted and not self.db.global.enable.override then
+        self.checkButtonAchievement:Hide()
+        self.container:SetPoint('BOTTOM', 0, -55)
+        self.container:SetSize(306, 50)
+        self.checkButtonKeystone:SetPoint('CENTER', -85, 0)
+    else
+        self.container:SetPoint('BOTTOM', 0, -75)
+        self.container:SetSize(306, 70)
+        self.checkButtonAchievement:SetPoint('CENTER', -85, 9)
+    end
+
+    self.checkButtonKeystone:Show()
 end